+++ src/General_Spotify_Helpers.py
@@ -20,7 +20,8 @@ import time
 from datetime import datetime, timedelta
 from typing import Optional
 
-from Settings import Settings
+from src.helpers.decorators import *
+from src.helpers.Settings import Settings
           
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Validates that the given 'args' are of type 'types'.
