+++ src/Spotify_Features.py
@@ -49,18 +49,18 @@ from datetime import datetime, timedelta
 from glob import glob
 from typing import Union
 
-import General_Spotify_Helpers as gsh
-from decorators import *
-from Settings import Settings
+import src.General_Spotify_Helpers as gsh
+from src.helpers.decorators import *
+from src.helpers.Settings import Settings
 
 # FEATURES
-from Misc_Features import MiscFeatures
-from Backup_Spotify_Data import BackupSpotifyData
-from Google_Drive_Uploader import DriveUploader
-from Log_Playback import LogPlayback
-from Sanity_Tests import SanityTest
-from Shuffle_Styles import Shuffler, ShuffleType
-from Weekly_Report import WeeklyReport
+from src.features.Misc_Features         import MiscFeatures
+from src.features.Backup_Spotify_Data   import BackupSpotifyData
+from src.features.Google_Drive_Uploader import DriveUploader
+from src.features.Log_Playback          import LogPlayback
+from src.features.Sanity_Tests          import SanityTest
+from src.features.Shuffle_Styles        import Shuffler, ShuffleType
+from src.features.Weekly_Report         import WeeklyReport
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Collection of all of our Spotify API features. Handles and abstracts our GSH object.
