+++ src/features/Google_Drive_Uploader.py
@@ -17,8 +17,8 @@ import logging
 from pydrive.auth import GoogleAuth
 from pydrive.drive import GoogleDrive
 
-from decorators import *
-from Settings import Settings
+from src.helpers.decorators import *
+from src.helpers.Settings import Settings
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Abstracted google drive api handler to upload single 'simple' (< 5MB) files.
