+++ src/helpers/decorators.py
@@ -12,7 +12,7 @@ import functools
 import logging
 from typing import Union, Optional
 
-from Settings import Settings
+from src.helpers.Settings import Settings
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Crates a simple Logger object with our desired format and returns it.
