+++ tests/run_unit_tests.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-export PYTHONPATH=$(pwd)/src:$PYTHONPATH
+export PYTHONPATH=$(pwd)/src:$(pwd)/tests:$PYTHONPATH
 source .venv/bin/activate
 pytest tests "$@"
