+++ tests/test_Backup_Spotify_Data.py
@@ -13,13 +13,14 @@ import sys
 import unittest
 
 # Override 'spotipy' with our local 'mocked_spotipy.py' MUST BE DONE BEFORE GSH
-sys.modules['spotipy'] = __import__('mocked_spotipy')
+from tests.helpers import mocked_spotipy
+sys.modules['spotipy'] = mocked_spotipy
 
-import General_Spotify_Helpers as gsh
-import tester_helpers as thelp
+import src.General_Spotify_Helpers as gsh
+import tests.helpers.tester_helpers as thelp
 
-from Backup_Spotify_Data import BackupSpotifyData, replace_none, get_column_types
-from Settings import Settings
+from src.features.Backup_Spotify_Data import BackupSpotifyData, replace_none, get_column_types
+from src.helpers.Settings import Settings
 
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 DESCRIPTION: Unit test collection for all Backup Spotify Data functionality.
